does not work

+++ packages/amberelec/package.mk
@@ -37,7 +37,7 @@ makeinstall_target() {
   mkdir -p $INSTALL/usr/config/
   rsync -av $PKG_DIR/config/* $INSTALL/usr/config/
 
-  if [ ! "$DEVICE" == "RG351MP" ] && [ ! "$DEVICE" == "RG351V" ]; then
+  if [ ! "$DEVICE" == "RG351MP" ] && [ ! "$DEVICE" == "RG351V" ]&& [ ! "$DEVICE" == "RG353P" ]; then
     rm -rf $INSTALL/usr/config/distribution/modules/display_fix.sh
   fi
 
@@ -48,6 +48,7 @@ makeinstall_target() {
     cp $INSTALL/usr/config/distribution/configs/distribution.conf.351p $INSTALL/usr/config/distribution/configs/distribution.conf
   elif  [ "$DEVICE" == "RG351V" ] || [ "$DEVICE" == "RG351MP" ]; then
     cp $INSTALL/usr/config/distribution/configs/distribution.conf.351v $INSTALL/usr/config/distribution/configs/distribution.conf
+  #FIXME: add RG353
   elif [ "$DEVICE" == "RG552" ]; then
     cp $INSTALL/usr/config/distribution/configs/distribution.conf.552  $INSTALL/usr/config/distribution/configs/distribution.conf
   fi
@@ -73,7 +74,7 @@ makeinstall_target() {
   mkdir -p $INSTALL/usr/share/retroarch-overlays
   if [ "$DEVICE" == "RG351P" ]; then
     cp -r $PKG_DIR/overlay-p/* $INSTALL/usr/share/retroarch-overlays
-  elif [ "$DEVICE" == "RG351V" ] || [ "$DEVICE" == "RG351MP" ]; then
+  elif [ "$DEVICE" == "RG351V" ] || [ "$DEVICE" == "RG351MP" ] || [ "$DEVICE" == "RG353P" ]; then
     cp -r $PKG_DIR/overlay-v/* $INSTALL/usr/share/retroarch-overlays
   elif [ "$DEVICE" == "RG552" ]; then
     cp -r $PKG_DIR/overlay-552/* $INSTALL/usr/share/retroarch-overlays
@@ -92,7 +93,7 @@ makeinstall_target() {
   mkdir -p $INSTALL/usr/share/bootloader
   if [ "$DEVICE" == "RG351P" ]; then
     find_file_path "splash/splash-480.bmp" && cp ${FOUND_PATH} $INSTALL//usr/share/bootloader/logo.bmp
-  elif [ "$DEVICE" == "RG351V" ] || [ "$DEVICE" == "RG351MP" ] ; then
+  elif [ "$DEVICE" == "RG351V" ] || [ "$DEVICE" == "RG351MP" ] || [ "$DEVICE" == "RG353P" ]; then
     find_file_path "splash/splash-640.bmp" && cp ${FOUND_PATH} $INSTALL//usr/share/bootloader/logo.bmp
   elif [ "$DEVICE" == "RG552" ]; then
     find_file_path "splash/splash-1920.bmp" && cp ${FOUND_PATH} $INSTALL//usr/share/bootloader/logo.bmp
